Add files via upload
This is a simulated dice game that I have personally never heard of. There can be between 2 and 4 players. Each of them can specify whether or not they want to roll the dice, if they do and roll a 1 their turn is over. If they roll higher then a 1, their rolls are added to their score. You can stop rolling whenever you would like. Whoever gets to 100 first wins. If two people each roll to 100, then the higher score wins.