Add files via upload
This is a countdown application. I wanted an app that I can put an event into and then put the date of that event and then tell me how many days left until that event. This is a base application. I will work on this more as I go. I would like to add a GUI to this application once I get better at Python.